feat: add Apache Cassandra ETL and data model project
The purpose of the project is to create an Apache Cassandra database
which can run queries on song play data to answer the questions,
which were specified in advance.

The 'ETL-and-data-model.ipynb' file provides for the code and
commentary.

The source data (provided by Udacity) in separate CSV files is not
included in Git, only the composite file 'event_datafile_new.csv'.